cpt/requirements.txt
torch>=2.2.0
transformers>=4.46.0
accelerate>=0.29.0
datasets>=2.18.0
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0
matplotlib>=3.8.0
huggingface_hub>=0.22.0
